+++ lib/core/chip_multiselect_field.dart
@@ -47,7 +47,7 @@ final class ChipMultiselectField extends StatelessWidget {
       shape: RoundedRectangleBorder(
         borderRadius: BorderRadius.circular(5),
         side: BorderSide(
-          color: Colors.grey.withOpacity(0.3), // ignore: deprecated_member_use
+          color: Colors.grey.withValues(alpha: 255 * 0.3),
           style: BorderStyle.solid,
           width: 0.5,
           strokeAlign: 0.5,
@@ -55,11 +55,8 @@ final class ChipMultiselectField extends StatelessWidget {
       ),
       deleteIcon: Align(
         alignment: Alignment.center,
-        child: Icon(
-          Icons.close,
-          size: 15,
-          color: Colors.grey.withOpacity(0.5), // ignore: deprecated_member_use
-        ),
+        child: Icon(Icons.close,
+            size: 15, color: Colors.grey.withValues(alpha: 255 * 0.5)),
       ),
       onDeleted: onDeleted,
     );